+++ docs/changelog/changelog.rst
@@ -5,6 +5,7 @@ develop
 * New SuiteBuilderProfiler that facilitates faster suite generation by allowing columns to be profiled
 * Added two convenience methods to ExpectationSuite: get_table_expectations & get_column_expectations
 * Added optional profiler_configuration to DataContext.profile() and DataAsset.profile()
+* Added list_available_expectation_types() to DataAsset
 
 0.9.11
 -----------------